Done: Integrate Maps API for add vacancy for another company.

+++ wp-content/themes/recruitment-valley/modules/post-type/class-vacancy.php
@@ -8,6 +8,7 @@ use Vacancy\Vacancy as VacancyModel;
 use constant\NotificationConstant;
 use Global\NotificationService;
 use Global\OptionController;
+use Helper\Maphelper;
 
 class Vacancy extends RegisterCPT
 {
@@ -19,6 +20,7 @@ class Vacancy extends RegisterCPT
     public function __construct()
     {
         add_action('init', [$this, 'RegisterVacancyCPT']);
+        add_action('save_post', [$this, 'vacancySubmitHandle'], 10, 3);
         add_action('set_object_terms', [$this, 'setExpiredDate'], 10, 5);
         add_action('add_meta_boxes', [$this, 'addVacancyMetaboxes'], 10, 2);
         add_filter('manage_vacancy_posts_columns', [$this, 'vacancyColoumn'], 10, 1);
@@ -250,11 +252,11 @@ class Vacancy extends RegisterCPT
         switch ($coloumn) {
             case 'status':
                 $status = $vacancyModel->getStatus();
-                if ($status['slug'] == 'open') {
+                if (array_key_exists('slug', $status) && $status['slug'] == 'open') {
                     echo '<span style="color: green; font-weight: bold;">' . $status['name'] . '<span>';
-                } else if ($status['slug'] == 'close') {
+                } else if (array_key_exists('slug', $status) && $status['slug'] == 'close') {
                     echo '<span style="color: red; font-weight: bold;">' . $status['name'] . '<span>';
-                } else if ($status['slug'] == 'declined') {
+                } else if (array_key_exists('slug', $status) && $status['slug'] == 'declined') {
                     echo '<span style="color: orange; font-weight: bold;">' . $status['name'] . '<span>';
                 } else {
                     echo '<span style="color: black; font-weight: bold;">' . $status['name'] . '<span>';
@@ -274,6 +276,134 @@ class Vacancy extends RegisterCPT
             case 'approvedat':
                 echo $vacancyModel->getApprovedAt('d M Y H:i:s');
                 break;
+        }
+    }
+
+    public function vacancySubmitHandle($post_id, $post, $update)
+    {
+        if ($post->post_type == 'vacancy') {
+            $this->wpdb->query('START TRANSACTION');
+            try {
+                $vacancyModel = new VacancyModel($post_id);
+                $vacancyData = [];
+
+                /** Calculate Vacancy City Coordiante */
+                // Check if city is not empty,
+                $vacancyData['city'] = $vacancyModel->getCity();
+                $vacancyData['city_latitude'] = $vacancyModel->getCityLongLat('latitude');
+                $vacancyData['city_longitude'] = $vacancyModel->getCityLongLat('longitude');
+                error_log(json_encode($vacancyData));
+
+                if ($vacancyData['city']) {
+                    if (!$vacancyData['city_longitude'] || empty($vacancyData['city_longitude']) || !$vacancyData['city_longitude'] || empty($vacancyData['city_longitude'])) {
+                        /** Calculate Coordinate */
+                        $vacancyModel->setCityLongLat($vacancyData['city']);
+                        $vacancyData['city_latitude'] = $vacancyModel->getCityLongLat('latitude');
+                        $vacancyData['city_longitude'] = $vacancyModel->getCityLongLat('longitude');
+                    }
+                }
+
+                /** Vacancy Placement Address Coordinatee */
+                $vacancyData['placement_address'] = $vacancyModel->getPlacementAddress();
+
+                /** Check if placement address is not empty */
+                if (isset($vacancyData['placement_address']) && !empty($vacancyData['placement_address'])) {
+                    $vacancyData['placement_address_latitude']   = $vacancyModel->getPlacementAddressLatitude();
+                    $vacancyData['placement_address_longitude']  = $vacancyModel->getPlacementAddressLongitude();
+                    if (!$vacancyData['placement_address_latitude'] || !$vacancyData['placement_address_longitude']) {
+                        /** Calculate Coordinate */
+                        $vacancyModel->setPlacementAddressLongitude($vacancyData['placement_address']);
+
+                        $vacancyData['placement_address_latitude']   = $vacancyModel->getPlacementAddressLatitude();
+                        $vacancyData['placement_address_longitude']  = $vacancyModel->getPlacementAddressLongitude();
+                    }
+                }
+
+                /** Set Distance */
+                error_log('the distance : ' . json_encode($vacancyModel->getDistance()));
+                if (!$vacancyModel->getDistance()) {
+                    error_log($post_id . ' - no distance');
+                    if ($vacancyData['city_latitude'] && $vacancyData['city_longitude'] && $vacancyData['placement_address_latitude'] && $vacancyData['placement_address_longitude']) {
+                        error_log($post_id . ' - setCoordinateDistance');
+                        $setCD = $vacancyModel->setCoordinateDistance([
+                            'lat' => $vacancyData['city_latitude'],
+                            'long' => $vacancyData['city_longitude']
+                        ], [
+                            'lat' => $vacancyData['placement_address_latitude'],
+                            'long' => $vacancyData['placement_address_longitude']
+                        ]);
+                        error_log($post_id . ' - setCoordinateDistance. ' . json_encode($setCD));
+                    } else if ($vacancyData['city'] && $vacancyData['placement_address']) {
+                        error_log($post_id . ' - setDistance');
+                        $setCD = $vacancyModel->setDistance($vacancyData["placement_city"], $vacancyData["placement_city"] . " " . $vacancyData["placement_address"]);
+                        error_log($post_id . ' - setDistance. ' . json_encode($setCD));
+                    }
+                }
+
+                /** Custom Company Address Coordinate */
+                /** Check if is_for_another_company */
+                if ($vacancyModel->checkIsForAnotherCompany()) {
+                    /** Check if not use_existing_company */
+                    if (!$vacancyModel->checkUseExistingCompany()) {
+                        $companyData['address_longitude'] = $vacancyModel->getCustomCompanyCoordinate('longitude');
+                        $companyData['address_latitude'] = $vacancyModel->getCustomCompanyCoordinate('latitude');
+                        $companyData['address'] = $vacancyModel->getCustomCompanyAddress();
+
+                        if (!$companyData['address_longitude'] || !$companyData['address_latitude']) {
+                            if ($companyData['address']) {
+                                $coordinat = Maphelper::generateLongLat($companyData['address']);
+                                $vacancyModel->setCustomCompanyLatitude($coordinat["lat"]);
+                                $vacancyModel->setCustomCompanyLongitude($coordinat["long"]);
+                            }
+                        }
+                    }
+                }
+
+                $this->wpdb->query('COMMIT');
+
+                /** Set status
+                 * IF free vacancy :
+                 *   - if new vacancy & if no status : set to processing.
+                 * IF paid vacancy :
+                 *   - if new vacancy & if no status : set to open.
+                 *
+                 * IF UPDATE then skip.
+                 *  */
+                if (!$update || !wp_is_post_revision($post)) {
+                    if ($vacancyModel->getIsPaid()) {
+                        /** if vacancy has no status, set to open */
+                        $vacancyStatus = $vacancyModel->getStatus();
+                        if (isset($vacancyStatus) && is_array($vacancyStatus)) {
+                            if (!empty($vacancyStatus['name'])) {
+                                if ($vacancyStatus['name'] !== 'open') {
+                                    $test = $vacancyModel->setStatus('open');
+                                }
+                            } else {
+                                $test = $vacancyModel->setStatus('open');
+                            }
+                        } else {
+                            $vacancyModel->setStatus('open');
+                        }
+                    } else {
+                        $vacancyStatus = $vacancyModel->getStatus();
+
+                        if (isset($vacancyStatus) && is_array($vacancyStatus)) {
+                            if (!empty($vacancyStatus['name'])) {
+                                if ($vacancyStatus['name'] !== 'processing') {
+                                    $vacancyModel->setStatus('processing');
+                                }
+                            } else {
+                                $vacancyModel->setStatus('processing');
+                            }
+                        } else {
+                            $vacancyModel->setStatus('processing');
+                        }
+                    }
+                }
+            } catch (\Exception $e) {
+                $this->wpdb->query('ROLLBACK');
+                error_log($e->getMessage());
+            }
         }
     }
 }
